Use setup_test.py as the new setup.py
